% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_cluster.R
\name{kmeans_cluster}
\alias{kmeans_cluster}
\title{Run k-means clustering}
\usage{
kmeans_cluster(data, xvar, yvar, k)
}
\arguments{
\item{data}{A data.frame}

\item{xvar}{Character name of x variable}

\item{yvar}{Character name of y variable}

\item{k}{Number of clusters}
}
\value{
A kmeans object
}
\description{
Run k-means clustering
}
