% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_clusters.R
\name{plot_clusters}
\alias{plot_clusters}
\title{Plot k-means clustering results}
\usage{
plot_clusters(data, xvar, yvar, fit)
}
\arguments{
\item{data}{A data.frame}

\item{xvar}{Character name of x variable}

\item{yvar}{Character name of y variable}

\item{fit}{A kmeans object (from kmeans_cluster)}
}
\value{
Invisibly returns NULL (creates a plot)
}
\description{
Plot k-means clustering results
}
